backend/pyproject.toml
[project]
name = "vibecon"
version = "0.1.0"
description = ""
authors = [
    {name = "Anand Krishnakumar ",email = "anandkrishnakumar123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "spotipy (>=2.25.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
